+++ src/Console/Application.php
@@ -9,17 +9,20 @@ use Castor\ContextDescriptor;
 use Castor\ContextRegistry;
 use Castor\Event\AfterApplicationInitializationEvent;
 use Castor\EventDispatcher;
+use Castor\ExpressionLanguage;
+use Castor\Fingerprint\FingerprintHelper;
 use Castor\FunctionFinder;
 use Castor\GlobalHelper;
 use Castor\ListenerDescriptor;
-use Castor\Monolog\Processor\ProcessProcessor;
 use Castor\PlatformUtil;
 use Castor\SectionOutput;
 use Castor\Stub\StubsGenerator;
 use Castor\TaskDescriptor;
 use Castor\VerbosityLevel;
+use Castor\WaitForHelper;
 use JoliCode\PhpOsHelper\OsHelper;
 use Monolog\Logger;
+use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Bridge\Monolog\Handler\ConsoleHandler;
 use Symfony\Component\Console\Application as SymfonyApplication;
 use Symfony\Component\Console\Command\Command;
@@ -28,13 +31,13 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\Filesystem\Filesystem;
+use Symfony\Component\Process\Process;
 use Symfony\Component\VarDumper\Cloner\AbstractCloner;
+use Symfony\Contracts\Cache\CacheInterface;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Contracts\HttpClient\Exception\ExceptionInterface as HttpExceptionInterface;
-
-use function Castor\log;
-use function Castor\request;
-use function Castor\run;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /** @internal */
 class Application extends SymfonyApplication
@@ -42,53 +45,88 @@ class Application extends SymfonyApplication
     public const NAME = 'castor';
     public const VERSION = 'v0.10.0';
 
+    // "Current" objects availables at some point of the lifecycle
+    private InputInterface $input;
+    private SectionOutput $sectionOutput;
+    private SymfonyStyle $symfonyStyle;
+    private Command $command;
+
     public function __construct(
         private readonly string $rootDir,
-        private readonly ContextRegistry $contextRegistry = new ContextRegistry(),
-        private readonly StubsGenerator $stubsGenerator = new StubsGenerator(),
-        private readonly FunctionFinder $functionFinder = new FunctionFinder(),
-        private readonly EventDispatcher $eventDispatcher = new EventDispatcher(),
+        public readonly FunctionFinder $functionFinder,
+        public readonly ContextRegistry $contextRegistry,
+        public readonly EventDispatcher $eventDispatcher,
+        public readonly ExpressionLanguage $expressionLanguage,
+        public readonly StubsGenerator $stubsGenerator,
+        public readonly Logger $logger,
+        public readonly Filesystem $fs,
+        public HttpClientInterface $httpClient,
+        public CacheItemPoolInterface&CacheInterface $cache,
+        public WaitForHelper $waitForHelper,
+        public FingerprintHelper $fingerprintHelper,
     ) {
-        if (!class_exists(\RepackedApplication::class)) {
-            $this->add(new RepackCommand());
-        }
-
         $this->setCatchErrors(true);
 
         AbstractCloner::$defaultCasters[self::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
         AbstractCloner::$defaultCasters[AfterApplicationInitializationEvent::class] = ['Symfony\Component\VarDumper\Caster\StubCaster', 'cutInternals'];
 
+        if (!class_exists(\RepackedApplication::class)) {
+            $this->add(new RepackCommand());
+        }
+
         parent::__construct(static::NAME, static::VERSION);
+
+        GlobalHelper::setApplication($this);
+    }
+
+    public function getInput(): InputInterface
+    {
+        return $this->input ?? throw new \LogicException('Input not available yet.');
+    }
+
+    public function getSectionOutput(): SectionOutput
+    {
+        return $this->sectionOutput ?? throw new \LogicException('Section output not available yet.');
+    }
+
+    public function getOutput(): OutputInterface
+    {
+        return $this->getSectionOutput()->getConsoleOutput();
+    }
+
+    public function getSymfonyStyle(): SymfonyStyle
+    {
+        return $this->symfonyStyle ?? throw new \LogicException('SymfonyStyle not available yet.');
+    }
+
+    public function getCommand(): Command
+    {
+        return $this->command ?? throw new \LogicException('Command not available yet.');
     }
 
     // We do all the logic as late as possible to ensure the exception handler
     // is registered
     public function doRun(InputInterface $input, OutputInterface $output): int
     {
-        $sectionOutput = new SectionOutput($output);
-
-        GlobalHelper::setApplication($this);
-        GlobalHelper::setEventDispatcher($this->eventDispatcher);
-        GlobalHelper::setInput($input);
-        GlobalHelper::setSectionOutput($sectionOutput);
-        GlobalHelper::setLogger(new Logger(
-            'castor',
-            [
-                new ConsoleHandler($sectionOutput->getConsoleOutput()),
-            ],
-            [
-                new ProcessProcessor(),
-            ]
-        ));
-        GlobalHelper::setContextRegistry($this->contextRegistry);
-        GlobalHelper::setupDefaultCache();
+        $this->input = $input;
+        $this->sectionOutput = new SectionOutput($output);
+        $this->symfonyStyle = new SymfonyStyle($input, $output);
+        $this->logger->pushHandler(new ConsoleHandler($output));
 
         $tasks = $this->initializeApplication($input);
 
-        GlobalHelper::setInitialContext($this->createContext($input, $output));
+        // Must be done after the initializeApplication() call, to ensure all
+        // contexts have been created; but before the adding of task, because we
+        // may want to seek in the context to know if the command is enabled
+        $this->contextRegistry->setCurrentContext($this->createContext($input, $output));
 
         foreach ($tasks as $task) {
-            $this->add(new TaskCommand($task->taskAttribute, $task->function));
+            $this->add(new TaskCommand(
+                $task->taskAttribute,
+                $task->function,
+                $this->eventDispatcher,
+                $this->expressionLanguage,
+            ));
         }
 
         return parent::doRun($input, $output);
@@ -96,7 +134,7 @@ class Application extends SymfonyApplication
 
     protected function doRunCommand(Command $command, InputInterface $input, OutputInterface $output): int
     {
-        GlobalHelper::setCommand($command);
+        $this->command = $command;
 
         if ('_complete' !== $command->getName() && !class_exists(\RepackedApplication::class)) {
             $this->stubsGenerator->generateStubsIfNeeded($this->rootDir . '/.castor.stub.php');
@@ -122,7 +160,7 @@ class Application extends SymfonyApplication
             if ($function instanceof TaskDescriptor) {
                 $tasks[] = $function;
             } elseif ($function instanceof ContextDescriptor) {
-                $this->contextRegistry->add($function);
+                $this->contextRegistry->addDescriptor($function);
             } elseif ($function instanceof ListenerDescriptor && null !== $function->reflectionFunction->getClosure()) {
                 $this->eventDispatcher->addListener(
                     $function->asListener->event,
@@ -137,7 +175,7 @@ class Application extends SymfonyApplication
         $contextNames = $this->contextRegistry->getNames();
 
         if ($contextNames) {
-            $defaultContext = PlatformUtil::getEnv('CASTOR_CONTEXT') ?: $this->contextRegistry->getDefault();
+            $defaultContext = PlatformUtil::getEnv('CASTOR_CONTEXT') ?: $this->contextRegistry->getDefaultName();
 
             $this->getDefinition()->addOption(new InputOption(
                 'context',
@@ -182,10 +220,10 @@ class Application extends SymfonyApplication
 
     private function displayUpdateWarningIfNeeded(SymfonyStyle $symfonyStyle): void
     {
-        $latestVersion = GlobalHelper::getCache()->get('latest-version', function (ItemInterface $item): array {
+        $latestVersion = $this->cache->get('latest-version', function (ItemInterface $item): array {
             $item->expiresAfter(3600 * 60 * 24);
 
-            $response = request('GET', 'https://api.github.com/repos/jolicode/castor/releases/latest', [
+            $response = $this->httpClient->request('GET', 'https://api.github.com/repos/jolicode/castor/releases/latest', [
                 'timeout' => 1,
             ]);
 
@@ -197,7 +235,7 @@ class Application extends SymfonyApplication
         });
 
         if (!$latestVersion) {
-            log('Failed to fetch latest Castor version from GitHub.');
+            $this->logger->info('Failed to fetch latest Castor version from GitHub.');
 
             return;
         }
@@ -217,7 +255,7 @@ class Application extends SymfonyApplication
             };
 
             if (!$assets) {
-                log('Failed to detect the correct release url adapted to your system.');
+                $this->logger->info('Failed to detect the correct release url adapted to your system.');
 
                 return;
             }
@@ -225,7 +263,7 @@ class Application extends SymfonyApplication
             $latestReleaseUrl = reset($assets)['browser_download_url'] ?? null;
 
             if (!$latestReleaseUrl) {
-                log('Failed to fetch latest phar url.');
+                $this->logger->info('Failed to fetch latest phar url.');
 
                 return;
             }
@@ -242,7 +280,9 @@ class Application extends SymfonyApplication
             return;
         }
 
-        $globalComposerPath = trim(run('composer global config home --quiet', quiet: true, allowFailure: true)->getOutput());
+        $process = new Process(['composer', 'global', 'config', 'home', '--quiet']);
+        $process->run();
+        $globalComposerPath = trim($process->getOutput());
 
         if ($globalComposerPath && str_contains(__FILE__, $globalComposerPath)) {
             $symfonyStyle->block('Run the following command to update Castor: <comment>composer global update jolicode/castor</comment>', escape: false);